fix(security): scope detect-secrets walk to prune venv/vendored dirs (#244)
detect-secrets was invoked with no scan path, so it walked the entire
working tree — venv/, node_modules/, large submodules — re-hashing
gigabytes. On big repos that sat right on the 60s timeout and made the
gate flakily "fail" with no actual finding. Its --exclude-files flag only
filters the report, not the walk, so a repo owner couldn't fix it; and
unlike bandit/semgrep, detect-secrets only accepts explicit paths.

Pass the top-level entries that aren't excluded as scan paths, reusing the
exact predicate that already post-filters findings — so the set of files
scanned (and secrets reported) is unchanged, only the wasted descent into
excluded dirs is removed (~49s -> ~1.7s in the reporter's measurements).

Adding the helper pushed security/__init__.py over the code-sprawl limit,
so the cohesive detect-secrets cluster moves to a DetectSecretsMixin in a
sibling module (verbatim; SecuritySubResult imported lazily to avoid an
import cycle, host members declared under TYPE_CHECKING).

Closes #244
