+++ tests/unit/test_detect_secrets_check.py
@@ -30,6 +30,59 @@ class TestRunDetectSecrets:
         assert check._is_path_excluded_for_detect_secrets("server/tests/test_auth.py")
         assert not check._is_path_excluded_for_detect_secrets("server/app/auth.py")
 
+    def test_scan_paths_prune_excluded_top_level_dirs(self, tmp_path):
+        """Helper drops venv/node_modules/dot-dirs but keeps real source.
+
+        Regression for barnacle #244: detect-secrets must not descend into
+        large vendored/venv directories — they cause the 60s-timeout flake.
+        """
+        (tmp_path / "src").mkdir()
+        (tmp_path / "venv").mkdir()
+        (tmp_path / "node_modules").mkdir()
+        (tmp_path / ".git").mkdir()
+        (tmp_path / "config.yml").write_text("name: ci\n")
+
+        check = SecurityLocalCheck({})
+        paths = check._detect_secrets_scan_paths(str(tmp_path))
+
+        assert "src" in paths
+        assert "config.yml" in paths
+        assert "venv" not in paths
+        assert "node_modules" not in paths
+        assert ".git" not in paths
+
+    def test_scan_paths_empty_when_root_unlistable(self):
+        """Unlistable root falls back to whole-tree scan (no path args)."""
+        check = SecurityLocalCheck({})
+        assert check._detect_secrets_scan_paths("/nonexistent/path/xyz") == []
+
+    def test_detect_secrets_scopes_walk_to_unexcluded_paths(self, tmp_path):
+        """_run_detect_secrets passes scoped paths so the walk is pruned.
+
+        The big vendored dirs must never reach the scan argv; the real source
+        dirs must. This is what takes the scan from ~49s to ~1.7s (#244).
+        """
+        (tmp_path / "src").mkdir()
+        (tmp_path / "venv").mkdir()
+        (tmp_path / "node_modules").mkdir()
+        (tmp_path / "config.yml").write_text("name: ci\n")
+
+        check = SecurityLocalCheck({})
+        mock_result = MagicMock()
+        mock_result.success = True
+        mock_result.output = json.dumps({"results": {}})
+
+        with patch.object(check, "_run_command", return_value=mock_result) as mock_run:
+            check._run_detect_secrets(str(tmp_path))
+
+        argv = mock_run.call_args[0][0]
+        assert "scan" in argv
+        assert "src" in argv
+        assert "config.yml" in argv
+        # The expensive vendored dirs must be pruned before the walk.
+        assert "venv" not in argv
+        assert "node_modules" not in argv
+
     def test_detect_secrets_no_findings(self, tmp_path):
         """Test _run_detect_secrets with no secrets found."""
         check = SecurityLocalCheck({})
@@ -650,3 +703,127 @@ class TestRunDetectSecrets:
             result = check._run_detect_secrets(str(tmp_path))
 
         assert result.passed is True
+
+
+class TestDetectSecretsHeuristics:
+    """Cover the detect-secrets post-processing helpers (moved to the mixin)."""
+
+    def test_false_positive_slopmop_path(self, tmp_path):
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path), ".slopmop/x.json", {"type": "Hex High Entropy String"}
+        )
+
+    def test_false_positive_flutter_ephemeral(self, tmp_path):
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path),
+            "ios/Flutter/ephemeral/Flutter-Generated.xcconfig",
+            {"type": "Base64 High Entropy String"},
+        )
+
+    def test_false_positive_xcscheme_hex(self, tmp_path):
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path), "App.xcscheme", {"type": "Hex High Entropy String"}
+        )
+
+    def test_false_positive_metadata_hex(self, tmp_path):
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path), "dir/.metadata", {"type": "Hex High Entropy String"}
+        )
+
+    def test_false_positive_env_example_basic_auth(self, tmp_path):
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path), ".env.example", {"type": "Basic Auth Credentials"}
+        )
+
+    def test_false_positive_hex_public_id_token(self, tmp_path):
+        (tmp_path / "cfg.py").write_text("account_id = 'deadbeefcafe1234'\n")
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path),
+            "cfg.py",
+            {"type": "Hex High Entropy String", "line_number": 1},
+        )
+
+    def test_false_positive_hex_git_sha_context(self, tmp_path):
+        (tmp_path / "g.py").write_text(
+            "ref = run(['git', 'rev-parse', 'HEAD'])  # abcdef1234567\n"
+        )
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path), "g.py", {"type": "Hex High Entropy String", "line_number": 1}
+        )
+
+    def test_false_positive_secret_keyword_getenv(self, tmp_path):
+        (tmp_path / "s.py").write_text("api_key = os.getenv('API_KEY')\n")
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path), "s.py", {"type": "Secret Keyword", "line_number": 1}
+        )
+
+    def test_false_positive_secret_keyword_placeholder(self, tmp_path):
+        (tmp_path / "t.py").write_text("password = 'change-me'\n")
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path), "t.py", {"type": "Secret Keyword", "line_number": 1}
+        )
+
+    def test_false_positive_secret_keyword_test_token(self, tmp_path):
+        (tmp_path / "u.py").write_text("test_secret = 'whatever-here'\n")
+        check = SecurityLocalCheck({})
+        assert check._is_detect_secrets_false_positive(
+            str(tmp_path), "u.py", {"type": "Secret Keyword", "line_number": 1}
+        )
+
+    def test_real_secret_keyword_is_not_false_positive(self, tmp_path):
+        (tmp_path / "r.py").write_text("password = 'aZ9RealLookingValue'\n")
+        check = SecurityLocalCheck({})
+        assert not check._is_detect_secrets_false_positive(
+            str(tmp_path), "r.py", {"type": "Secret Keyword", "line_number": 1}
+        )
+
+    def test_safe_read_line_invalid_line_number(self, tmp_path):
+        check = SecurityLocalCheck({})
+        assert check._safe_read_line(str(tmp_path), "x.py", None) == ""
+        assert check._safe_read_line(str(tmp_path), "x.py", 0) == ""
+
+    def test_safe_read_line_reads_and_caches(self, tmp_path):
+        (tmp_path / "f.py").write_text("line1\nline2\n")
+        check = SecurityLocalCheck({})
+        cache: dict = {}
+        assert check._safe_read_line(str(tmp_path), "f.py", 2, cache) == "line2"
+        # Second read hits the cache branch.
+        assert check._safe_read_line(str(tmp_path), "f.py", 1, cache) == "line1"
+
+    def test_create_plugin_config_baseline_none_without_config(self, tmp_path):
+        check = SecurityLocalCheck({})
+        assert check._create_plugin_config_baseline(str(tmp_path)) is None
+
+    def test_create_plugin_config_baseline_writes_throwaway(self, tmp_path):
+        (tmp_path / ".secrets.baseline").write_text(
+            json.dumps(
+                {
+                    "plugins_used": [{"name": "Base64HighEntropyString"}],
+                    "version": "1.4.0",
+                }
+            )
+        )
+        check = SecurityLocalCheck({"config_file_path": ".secrets.baseline"})
+        path = check._create_plugin_config_baseline(str(tmp_path))
+        assert path is not None
+        data = json.loads(Path(path).read_text())
+        assert data["plugins_used"] == [{"name": "Base64HighEntropyString"}]
+        assert data["results"] == {}
+        Path(path).unlink()
+
+    def test_load_allowlist_returns_known_pairs(self, tmp_path):
+        (tmp_path / ".secrets.baseline").write_text(
+            json.dumps({"results": {"a.py": [{"hashed_secret": "abc123"}]}})
+        )
+        check = SecurityLocalCheck({"config_file_path": ".secrets.baseline"})
+        known = check._load_detect_secrets_allowlist(str(tmp_path))
+        assert ("a.py", "abc123") in known
